[Blog] New functional tests

+++ src/Services/Blog/Blog.API/Services/BlogService.cs
@@ -45,30 +45,46 @@ namespace Blog.API.Services
         {
             var blog = await _repository.GetByIdAsync(blogId);
 
-            if (blog == null || blog.SoftDeletedAt.HasValue) return new NotFoundResult();
+            if (blog == null || blog.SoftDeletedAt.HasValue)
+            {
+                return new NotFoundResult();
+            }
+
+            // TODO: Check for owner existence with IdentityControl
 
             // For now. Will be changed after creating AccessControl
-            if (blog.OwnerId != _userInfo.Id && _userInfo.Role != Roles.Administrator) return new ForbidResult();
+            if (blog.OwnerId != _userInfo.Id && _userInfo.Role != Roles.Administrator)
+            {
+                return new ForbidResult();
+            }
 
             blog.OwnerId = newOwnerId;
-            blog.UpdatedAt = DateTime.UtcNow;
             await _repository.UpdateAsync(blog);
 
             return new OkResult();
         }
 
-        public async Task<IActionResult> DisableAsync(string blogId)
+        public async Task<IActionResult> ToggleEnabledState(string blogId, bool enable)
         {
             var blog = await _repository.GetByIdAsync(blogId);
 
-            if (blog == null) return new NotFoundResult();
+            if (blog == null)
+            {
+                return new NotFoundResult();
+            }
 
-            if (blog.DisabledAt.HasValue) return new BadRequestResult();
+            if ((enable && !blog.DisabledAt.HasValue) || (!enable && blog.DisabledAt.HasValue))
+            {
+                return new BadRequestResult();
+            }
 
             // For now. Will be changed after creating AccessControl
-            if (blog.OwnerId != _userInfo.Id && _userInfo.Role != Roles.Administrator) return new ForbidResult();
+            if (blog.OwnerId != _userInfo.Id && _userInfo.Role != Roles.Administrator)
+            {
+                return new ForbidResult();
+            }
 
-            blog.DisabledAt = DateTime.UtcNow;
+            blog.DisabledAt = enable ? null : DateTime.UtcNow;
             await _repository.UpdateAsync(blog);
 
             return new OkResult();
@@ -78,7 +94,10 @@ namespace Blog.API.Services
         {
             var blog = await _repository.GetByIdAsync(id);
 
-            if (blog != null) return new OkObjectResult(blog);
+            if (blog != null)
+            {
+                return new OkObjectResult(blog);
+            }
 
             return new NotFoundResult();
         }
@@ -90,7 +109,10 @@ namespace Blog.API.Services
 
         public async Task<ActionResult<List<BlogModel>>> GetAllAsync()
         {
-            if (_userInfo.Role != Roles.Administrator) return new ForbidResult();
+            if (_userInfo.Role != Roles.Administrator)
+            {
+                return new ForbidResult();
+            }
 
             var blogs = await _repository.Query().Find(x => !x.SoftDeletedAt.HasValue).ToListAsync();
             return new OkObjectResult(blogs);
@@ -98,28 +120,40 @@ namespace Blog.API.Services
 
         public async Task<IActionResult> UpdateAsync(BlogUpdateDto blog)
         {
-            var entity = new BlogModel
-            {
-                Id = blog.Id,
-                Name = blog.Name,
-                Title = blog.Title,
-                Heading = blog.Heading,
-                Footer = blog.Footer,
-                Path = blog.Path,
-                Uri = blog.Uri,
-                UpdatedAt = DateTime.UtcNow
-            };
+            var entity = await _repository.GetByIdAsync(blog.Id);
+
+            entity.Name = blog.Name;
+            entity.Title = blog.Title;
+            entity.Heading = blog.Heading;
+            entity.Footer = blog.Footer;
+            entity.Path = blog.Path;
+            entity.Uri = blog.Uri;
+            entity.UpdatedAt = DateTime.UtcNow;
+
             var result = await _repository.UpdateAsync(entity);
-            if (result.IsModifiedCountAvailable && result.ModifiedCount == 1) return new OkResult();
+            if (result.IsModifiedCountAvailable && result.ModifiedCount == 1)
+            {
+                return new OkResult();
+            }
 
             return new NotFoundResult();
         }
 
         public async Task<IActionResult> DeleteAsync(string blogId)
         {
+            var entity = await _repository.GetByIdAsync(blogId);
+
+            if (entity.OwnerId != _userInfo.Id)
+            {
+                return new ForbidResult();
+            }
+
             var result = await _repository.DeleteAsync(blogId);
 
-            if (result.DeletedCount == 1) return new NoContentResult();
+            if (result.DeletedCount == 1)
+            {
+                return new NoContentResult();
+            }
 
             return new NotFoundResult();
         }
@@ -127,10 +161,24 @@ namespace Blog.API.Services
         public async Task<IActionResult> SoftDeleteAsync(string blogId)
         {
             var entity = await _repository.GetByIdAsync(blogId);
+
+            if (entity == null)
+            {
+                return new NotFoundResult();
+            }
+
+            if (entity.OwnerId != _userInfo.Id && _userInfo.Role != Roles.Administrator)
+            {
+                return new ForbidResult();
+            }
+
             entity.SoftDeletedAt = DateTime.UtcNow;
             var result = await _repository.UpdateAsync(entity);
 
-            if (result.ModifiedCount == 1) return new OkResult();
+            if (result.ModifiedCount == 1)
+            {
+                return new OkResult();
+            }
 
             return new NotFoundResult();
         }
